backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy>=2.0.46
python-multipart==0.0.9
python-docx==1.1.0
jinja2==3.1.3
aiofiles==23.2.1
openai>=1.0.0
python-dotenv>=1.0.0
itsdangerous>=2.0.0
